feat(web): add Shell component with menu and color scheme support
This commit adds a new `Shell` component that wraps the entire app and provides a header, navbar, and color scheme support. The `Shell` component accepts a `menu` prop that can be used to render a navigation menu in the navbar. The `Shell` component also provides a `ShellContext` that can be used to update the menu from child components.

The commit also adds support for toggling between light and dark color