refactor(user): add completeGetUser flag to User context
Add a new flag `completeGetUser` to the `User` context to indicate whether the user data has been completely fetched or not. This flag is used in the `Profile` page to redirect the user to the login page if the user data is not available. Also, remove unnecessary code related to redirecting the user to the login page in the `Profile` page.